Database に検索結果を保存して画面に表示できるようにした

+++ app/src/main/java/com/example/arakitaku/aactwitterclient/RepositoryDao.java
@@ -15,7 +15,7 @@ import java.util.List;
 public interface RepositoryDao {
 
     @Query("SELECT * FROM Repository WHERE id in (:repositoryIds)")
-    LiveData<List<Repository>> getByIds(List<Long> repositoryIds);
+    LiveData<List<Repository>> getByIds(List<Integer> repositoryIds);
 
     @Query("SELECT * FROM SearchResult WHERE query = :query")
     LiveData<SearchResult> search(String query);
